move geohash and geoservies parsing functions
fixes to index creation

delinting
