+++ lib/table.js
@@ -1,5 +1,5 @@
 var _ = require('lodash')
-var Indexing = require('./indexing')
+var Indexes = require('./indexes')
 var Geohash = require('./geohash')
 
 var Table = {}
@@ -46,7 +46,7 @@ Table.createFeatureTable = function (id, geojson, layerId, callback) {
     feature = { geometry: { type: geojson.geomType || types[geojson.info.geometryType] } }
   }
 
-  var indexes = Indexing.prepareIndexes(info)
+  var indexes = Indexes.prepare(info)
 
   Table.create(table, buildSchemaFromFeature(feature), indexes, function (err) {
     if (err) {
@@ -94,7 +94,7 @@ Table.create = function (name, schema, indexes, callback) {
     self.query(create, function (err, result) {
       if (err) return callback('Failed to create table ' + name + ' error:' + err)
       if (!indexes || !indexes.length) return callback()
-      Indexing._addIndexes(name, indexes, function (err) {
+      Indexes._add(name, indexes, function (err) {
         callback(err)
       })
     })
@@ -155,14 +155,7 @@ function prepareFeature (feature) {
  */
 function buildSchemaFromFeature (feature) {
   var schema = '('
-  var type
-  if (feature && feature.geometry && feature.geometry.type) {
-    type = feature.geometry.type.toUpperCase()
-  } else {
-    // default to point geoms
-    type = 'POINT'
-  }
-  var props = ['id SERIAL PRIMARY KEY', 'feature JSON', 'geom Geometry(' + type + ', 4326)', 'geohash varchar(10)']
+  var props = ['id SERIAL PRIMARY KEY', 'feature JSON', 'geohash varchar(10)']
   schema += props.join(',') + ')'
   return schema
 }
